Add entry/exit events for module, verify debug works
* Installing the module
	* `make` (clean is only other option)
	* `sudo insmod fbi.ko`

* Uninstalling module
	* `sudo rmmod fbi`

* Looking at Kernel output (printk)
	* change filtering tolerances to allow KERN_NOTICE events
	* `cat /proc/sys/kernel/printk` will show default log levels
		* see proc(5) man pages
		* first entry by default is often 4, or KERN_WARNING
		* KERN_[NOTICE|INFO|DEBUG] will be filtered since ^

	* `sudo dmesg -n 7` will change the value of the first entry
	* Note: can automate log level into kernel boot params

* Note: currently getting a message:
"Module verfication failed: signature and/or required key missing -
tainting kernel."

While this does not disable loading the module into the kernel, I
think this is just designed to be explicit about the fact that we are
introducing new, unsigned source code into the kernel. We want to try
and remove this message if possible for the rootkit to be stealthy.